Chrome Mobile & New Projects Page
The background as well as the navbar are now both fixed for Chrome Mobile. To make room for this fix the navbar has been changed a little.
The projects page has also been updated with a whole new look. Grid-items now consist of the project image as well as a title and some text explaining what said projects is.

Pog